[WIP] test(VAutocomplete): convert tests to vue-test-utils (#6776)
* chore(VAutocomplete): move and rename tests

* test(VAutocomplete): convert to vue-test-utils

* style(VAutocomplete): fix linting errors

* test(VAutocomplete): fix more tests

* test(VAutocomplete): fix lint errors

* test(VAutocomplete): skip last test

will revisit when component gets refactored.
